API URL logic & ReuseableTableWithModal Updates
Changed the logic of the API url - moved it from being hard coded in individual pages, to being stored in a .env.local file. This file will be updated automatically by the web host when custom env's are saved. This will make the API point to the hosted backend when in prod, and use localhost when editing locally.

handleApiCall now automatically adds the auth token to calls and has predefined endpoints exported out for use in pages. Also allows a custom api call if a predefined endpoint isn't available.

ReuseableTableWithModal has been updated with the above logic. It also allows limited customization of the modal layout by allowing the user to pass the layout to the component itself.